+++ packages/button/src/types.ts
@@ -23,7 +23,7 @@ export const Size = {
   Small: 'small',
   Default: 'default',
   Large: 'large',
-};
+} as const;
 
 export type Size = (typeof Size)[keyof typeof Size];