+++ packages/menu/src/MenuItem/MenuItem.types.ts
@@ -1,16 +1,15 @@
 import { Size } from '../types';
 
+const Variant = {
+  Default: 'default',
+  Destructive: 'destructive',
+} as const;
+
+type Variant = (typeof Variant)[keyof typeof Variant];
+
+export { Variant };
+
 export interface MenuItemProps {
-  /**
-   * Determines whether or not the MenuItem is active.
-   */
-  active?: boolean;
-
-  /**
-   * Description element displayed below title in MenuItem.
-   */
-  description?: React.ReactNode;
-
   /**
    * Determines whether or not the MenuItem is disabled.
    */
@@ -31,6 +30,21 @@ export interface MenuItemProps {
    * Content to appear inside of `<MenuItem />` component
    */
   children?: React.ReactNode;
+
+  /**
+   * Determines whether or not the MenuItem is active.
+   */
+  active?: boolean;
+
+  /**
+   * Description element displayed below title in MenuItem.
+   */
+  description?: React.ReactNode;
+
+  /**
+   * Variant of MenuItem
+   */
+  variant?: Variant;
 }
 
 export interface FocusableMenuItemProps {