+++ packages/search-input/src/SearchResultsMenu/SearchResultsMenu.types.ts
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { ReactElement } from 'react';
 
 import { HTMLElementProps } from '@leafygreen-ui/lib';
 import { PortalControlProps } from '@leafygreen-ui/popover';
@@ -7,4 +7,5 @@ export type SearchResultsMenuProps = HTMLElementProps<'ul', HTMLUListElement> &
   PortalControlProps & {
     refEl: React.RefObject<HTMLElement>;
     open?: boolean;
+    footerSlot?: ReactElement;
   };