+++ packages/input-option/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @leafygreen-ui/input-option
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies [dd4f3da8]
+- Updated dependencies [90053e16]
+  - @leafygreen-ui/lib@13.0.0
+  - @leafygreen-ui/typography@18.0.0
+  - @leafygreen-ui/a11y@1.4.11
+  - @leafygreen-ui/leafygreen-provider@3.1.10
+
+## 1.0.12
+
+### Patch Changes
+
+- c5d95794: Exports `descriptionClassName` and adds `font-size: inherit;` to input descriptions.
+- Updated dependencies [324d9730]
+  - @leafygreen-ui/typography@17.0.2
+
 ## 1.0.11
 
 ### Patch Changes