+++ packages/search-input/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @leafygreen-ui/search-input
 
+## 2.1.2
+
+### Patch Changes
+
+- 9a471879: - Updates dark mode dropdown border from `gray.dark3` to `gray.dark2`
+- Updated dependencies [dd4f3da8]
+- Updated dependencies [784e9d8a]
+- Updated dependencies [90053e16]
+  - @leafygreen-ui/lib@13.0.0
+  - @leafygreen-ui/icon-button@15.0.19
+  - @leafygreen-ui/typography@18.0.0
+  - @leafygreen-ui/a11y@1.4.11
+  - @leafygreen-ui/input-option@1.0.13
+  - @leafygreen-ui/leafygreen-provider@3.1.10
+  - @leafygreen-ui/popover@11.1.1
+
+## 2.1.1
+
+### Patch Changes
+
+- d70768a7: Add a 1px `gray.light2` border to light mode dropdowns and a 1px `gray.dark3` border to dark mode dropdowns to match Figma specs.
+
+## 2.1.0
+
+### Minor Changes
+
+- f6698369: SearchResultMenu now supports a footer slot prop for children rendered under the list of search results. Popover props are also now correctly passed to the Popover component.
+
+### Patch Changes
+
+- Updated dependencies [f8c77c5d]
+- Updated dependencies [1c7c9608]
+- Updated dependencies [827060f6]
+  - @leafygreen-ui/popover@11.1.0
+  - @leafygreen-ui/icon@11.24.0
+
 ## 2.0.17
 
 ### Patch Changes